refactor: extract from root command to runx package
